Changing to no use any options as shell=true isn't working
Using shell seems to cause client.spec to not find the 'invokeNop'id
which causes a error to be thrown.

But there is a second problem, when I caught the error in the catch
block I execute an Expect(false).toBe(true) which fails but the
yarn travis exited with a 0 rather than a non-zero so the error wasn't
reported:
```

$ DEBUG=client.spec alsatian 'build/client/**/*.spec.js' --tap
  client.spec setupFixture:+ +0ms
TAP version 13
1..1
  client.spec server.stdout: /bin/sh: 1: node: not found
  client.spec  +33ms
  client.spec server: exited with 127 +2ms
  client.spec setupFixture:- +876ms
# FIXTURE Client tests
  client.spec testWdNop:+ +5ms
testWdNop caught e=NoSuchElementError: no such element: Unable to locate element: {"method":"css selector","selector":"*[id="invokeNop"]"}
  (Session info: chrome=57.0.2987.98)
  (Driver info: chromedriver=2.28.455506 (18f6627e265f442aeec9b6661a49fe819aeeea1f),platform=Linux 4.4.0-51-generic x86_64)
not ok 1 wd: nop which is defined in a TS file
 ---
   message: "Expected false to be true."
   severity: fail
   data:
     got: false
     expect: true
 ...
  client.spec teardownFixture:+ +39ms
  client.spec teardownFixture:- +56ms
The command "yarn test:travis" exited with 0.
```
I wonder what I've done wrong.