Update server.ts and server.spec.ts
Server writes `running PORT=${port}` to stdout when its listening so that
the test spec's may know when the server starts.

Added `/nop` handler to server so we can explicitly test nop in the server.

Server.spec now waits until the server has started before proceeding
and also does httpGet('/') of the root for the heck of it and
invokes httpGet('/nop') in testNop.

Remove testNothing.